ddsrotatingfilehandler moved to own file
