+++ dds_web/__init__.py
@@ -30,66 +30,6 @@ C_TZ = pytz.timezone("Europe/Stockholm")
 oauth = auth_flask_client.OAuth(app)
 actions = {"api_blueprint.auth": "User Authentication", "api_blueprint.proj_auth": "Project Access"}
 
-# CLASSES ########################################################### CLASSES #
-
-
-class DDSRotatingFileHandler(logging.handlers.RotatingFileHandler):
-    def __init__(
-        self,
-        filename,
-        basedir,
-        mode="a",
-        maxBytes=1e9,
-        backupCount=0,
-        encoding=None,
-        delay=0,
-    ):
-        """
-        Custom RotatingFileHandler, logs to the file `<basedir>/<filename>.log`
-        and renames the current file to `<basedir>/<filename>_[timestamp].log` when the file size
-        reaches <maxBytes> --> Current logging always to <filename>.log.
-        """
-
-        self.today_ = datetime.now() if not hasattr(self, "today_") else self.today_
-        self.basedir_ = pathlib.Path(basedir)  # Log directory
-        self.basename = pathlib.Path(filename)  # Base for all filenames
-        self.active_file_name = self.basedir_ / self.basename.with_suffix(".log")  # Active file
-
-        # Initiate super class
-        logging.handlers.RotatingFileHandler.__init__(
-            self, self.active_file_name, mode, maxBytes, backupCount, encoding, delay
-        )
-
-    def shouldRollover(self, record):
-        """
-        Checks if the FileHandler should do a rollover of the log file.
-        """
-
-        if self.stream is None:
-            self.stream = self._open()
-
-        # Check if the file is at max size
-        if self.maxBytes > 0:
-            msg = "%s\n" % self.format(record)
-            self.stream.seek(0, 2)
-            if self.stream.tell() + len(msg) >= self.maxBytes:
-                # Create time stamp and rename the current log file to contain rollover timestamp
-                new_today = datetime.now()
-                replacement_name = pathlib.Path(
-                    str(self.basename)
-                    + "_"
-                    + self.today_.strftime("%Y-%m-%d-%H-%M-%S")
-                    + "_"
-                    + new_today.strftime("%Y-%m-%d-%H-%M-%S")
-                    + ".log"
-                )
-                self.active_file_name.rename(target=pathlib.Path(self.basedir_ / replacement_name))
-                self.today_ = new_today
-                return 1
-
-        return 0
-
-
 # FUNCTIONS ####################################################### FUNCTIONS #
 
 
@@ -128,14 +68,14 @@ def create_app():
             "handlers": {
                 "general": {
                     "level": logging.DEBUG,
-                    "class": "dds_web.DDSRotatingFileHandler",
+                    "class": "dds_web.dds_rotating_file_handler.DDSRotatingFileHandler",
                     "filename": "dds",
                     "basedir": app.config.get("LOG_DIR"),
                     "formatter": "general",
                 },
                 "actions": {
                     "level": logging.INFO,
-                    "class": "dds_web.DDSRotatingFileHandler",
+                    "class": "dds_web.dds_rotating_file_handler.DDSRotatingFileHandler",
                     "filename": "actions",
                     "basedir": app.config.get("LOG_DIR"),
                     "formatter": "actions",
